Add Node.js project setup with dependencies and Cypress configuration
This commit introduces a Node.js project setup with the addition of
`package.json`, which specifies project metadata and dependencies,
including `cherry-cljs` and `devDependencies` like `cypress`. A
`package-lock.json` is generated to lock down specific versions of
these dependencies for consistent installations. The `.gitignore` file
is updated to ignore `node_modules`.

Additionally, `shell.nix` is updated to configure Nix development
environments, now including Node.js and a specific version of
`cypress` that is fetched from an external URL. The
`CYPRESS_INSTALL_BINARY` and `CYPRESS_RUN_BINARY` environment
variables are set to control Cypress binary installations and usage in
a Unix environment. This setup prepares the project for both Node
development and testing workflows utilizing Cypress.